XHCI: Increase fScratchpadCount size
* To fix clang's '-Wtautological-constant-out-of-range-compare' warnings in xhci.cpp.
* by korli: adjust trace print macro.

+++ src/add-ons/kernel/busses/usb/xhci.h
@@ -242,7 +242,7 @@ private:
 			uint8				fPortSlots[XHCI_MAX_PORTS];
 
 			// Scratchpad
-			uint8				fScratchpadCount;
+			uint32				fScratchpadCount;
 			area_id				fScratchpadArea[XHCI_MAX_SCRATCHPADS];
 			void *				fScratchpad[XHCI_MAX_SCRATCHPADS];
